Tadaa, ANOVA!

Performs one-, two-way or factorial ANOVA with adjustable sums of squares method and optionally displays effect sizes ((partial) \(\eta^2\), Cohen's f) and power (calculated via pwr.f2.test to work with unbalanced designs).

tadaa_aov(formula, data = NULL, show_effect_size = TRUE,
  show_power = TRUE, factorize = TRUE, type = 3, check_contrasts = TRUE,
  print = c("df", "console", "html", "markdown"))

Arguments

formula
Formula for model, passed to aov.

data
Data for model.

show_effect_size
If TRUE (default), effect sizes partial eta^2 and Cohen's f are appended as columns.

show_power
(Experimental) If TRUE (default), power is calculated via pwr.f2.test and appended as a column.

factorize
If TRUE (default), non-factor independent variables will automatically converted via as.factor, so beware of your inputs.

type
Which type of SS to use. Default is 3, can also be 1 or 2.

check_contrasts
Only applies to type = 3. If TRUE (default), the contrasts of each non-ordered factor are set to "contr.sum".

print
Print method, default df: A regular data.frame. Otherwise passed to sprinkle_print_method for fancyness.

Value

A data.frame by default, otherwise dust object, depending on print.

Details

If a specified independent variable is not properly encoded as a factor, it is automatically converted if factorize = TRUE to ensure valid results.

If type = 3 and check_contrasts = TRUE, the "contrasts" of each non-ordered factor will be checked and set to contr.sum to ensure the function yields usable results. It is highly recommended to only use check_contrasts = FALSE for debugging or educational purposes.

See also

Other Tadaa-functions: tadaa_chisq, tadaa_kruskal, tadaa_levene, tadaa_nom, tadaa_normtest, tadaa_one_sample, tadaa_ord, tadaa_pairwise_gh, tadaa_pairwise_tukey, tadaa_pairwise_t, tadaa_t.test, tadaa_wilcoxon

Examples

tadaa_aov(stunzahl ~ jahrgang, data = ngo)

#>        term  df   sumsq         meansq     statistic           p.value
#> 1  jahrgang   2  536.28 268.1400000000 26.4782514832 3.81395822294e-11
#> 2 Residuals 247 2501.32  10.1268016194            NA                NA
#> 3     Total 249 3037.60 278.2668016194            NA                NA
#>      eta.sq.part       cohens.f          power
#> 1 0.176547274164 0.463032177537 0.999999614919
#> 2             NA             NA             NA
#> 3             NA             NA             NA

tadaa_aov(stunzahl ~ jahrgang * geschl, data = ngo)

#>              term  df    sumsq          meansq       statistic
#> 1          geschl   1    7.290   7.29000000000  0.739572245875
#> 2        jahrgang   2  536.280 268.14000000000 27.202867216605
#> 3 jahrgang:geschl   2   96.056  48.02800000000  4.872452102182
#> 4       Residuals 244 2405.120   9.85704918033              NA
#> 5           Total 249 3044.746 333.31504918033              NA
#>             p.value      eta.sq.part        cohens.f          power
#> 1 3.90642106132e-01 0.00302187439117 0.0550548253523 0.138097664443
#> 2 2.16314047232e-11 0.18232134357789 0.4722015686628 0.999999768124
#> 3 8.41559122985e-03 0.03840433460100 0.1998452701094 0.804726341974
#> 4                NA               NA              NA             NA
#> 5                NA               NA              NA             NA

# Other types of sums and print options
not_run({
  tadaa_aov(stunzahl ~ jahrgang * geschl, data = ngo, type = 1, print = "console")
  tadaa_aov(stunzahl ~ jahrgang * geschl, data = ngo, type = 3, print = "console")
  tadaa_aov(stunzahl ~ jahrgang * geschl, data = ngo,
            type = 3, check_contrasts = FALSE, print = "console")
})