Tadaa, test for normality!

Tadaa, test for normality!

tadaa_normtest(data, method = "ad", print = c("df", "console", "html",
  "markdown"), ...)

Arguments

data
A data.frame.

method
The type of test to perform. Either ad for Anderson Darling, shapiro for Shapiro-Wilk, pearson for Pearson's chi-square test or ks for Kolmogorov-Smirnov (not recommended).

print
Print method, default df: A regular data.frame. Otherwise passed to pixiedust::sprinkle_print_method for fancyness.

...
Further arguments passed to test functions where applicable, see nortest::pearson.test and stats::ks.test.

Value

A data.frame by default, otherwise dust object, depending on print.

See also

Examples

# NOT RUN {
library(dplyr)
ngo %>%
  select(englisch, deutsch, mathe) %>%
  tadaa_normtest(method = "shapiro")

ngo %>%
  select(englisch, deutsch, mathe) %>%
  tadaa_normtest(method = "pearson", print = "console")
# }