Tadaa, one-sample tests!

If `sigma` is omitted, the function will basically just perform a one-sample t.test, but if `sigma` is provided, a z-test is performed. It basically works the same way, except that we pretend we know the population sigma.

tadaa_one_sample(data = NULL, x, mu, sigma = NULL,
  direction = "two.sided", na.rm = TRUE, print = "df")

Arguments

data
A data.frame.

x
A numeric vector.

mu
The true mean (\(\mu\)) to test for.

sigma
Population sigma. If supplied, a z-test is performed, else a one-sample t.test is performed.

direction
Test direction, like alternative in t.test.

na.rm
If TRUE (default), missing values are dropped.

print
Print method, per default a regular data.frame. Otherwise passed to sprinkle_print_method for fancyness.

Value

A data.frame by default, otherwise dust object, depending on print.

See also

Other Tadaa-functions: tadaa_aov, tadaa_kruskal, tadaa_nom, tadaa_normtest, tadaa_ord, tadaa_t.test, tadaa_wilcoxon

Examples

df <- data.frame(x = rnorm(n = 20, mean = 100, sd = 1))

tadaa_one_sample(df, x, mu = 101, sigma = 1)

#>   estimate statistic      p.value conf.low conf.high method alternative
#> 1 100.1102 -3.979176 6.915454e-05 80.64757  119.5729 z-test   two.sided
#>            d     power
#> 1 -0.8897708 0.9782674

# No data.frame, just a vector
tadaa_one_sample(x = rnorm(20), mu = 0)

#>     estimate statistic   p.value parameter  conf.low conf.high
#> 1 -0.2223487 -0.805793 0.4303323        19 -0.799893 0.3551956
#>              method alternative          d     power
#> 1 One Sample t-test   two.sided -0.1801808 0.1193946