Games Howell Post-Hoc Test

An implementation of the Games Howell procedure for pairwise comparisons. The workhorse of this function is adapted from this gist: https://gist.github.com/aschleg/ea7942efc6108aedfa9ec98aeb6c2096

tadaa_pairwise_gh(data, response, group1, group2 = NULL, print = "df")

Arguments

data
A data.frame containing the variables.

response
The response variable, i.e. the dependent numeric vector.

group1
The grouping variables, typically a factor.

group2
(Optional) second grouping variable.

print
Print method, defaults to df for data.frame output, otherwise passed to pixiedust::sprinkle_print_method.

Source

https://gist.github.com/aschleg/ea7942efc6108aedfa9ec98aeb6c2096

Value

A data.frame or pixiedust::dust object depending on print.

Note

This function is really, really slow for large comparisons (k > 50). Sorry about that.

References

https://rpubs.com/aaronsc32/games-howell-test

See also

Examples

tadaa_pairwise_gh(ngo, deutsch, jahrgang)

#>       term comparison    se     t      df conf_low conf_high adj.p.value
#> 1 jahrgang    11 : 12 0.200 0.176 193.089   -0.619     0.719       0.983
#> 2 jahrgang    11 : 13 0.253 3.276 104.415    0.321     2.019       0.004
#> 3 jahrgang    12 : 13 0.240 3.306  90.538    0.313     1.927       0.004

tadaa_pairwise_gh(ngo, deutsch, jahrgang, geschl)

#>               term                    comparison    se     t      df conf_low
#> 1         jahrgang                       11 : 12 0.200 0.176 193.089   -0.619
#> 2         jahrgang                       11 : 13 0.253 3.276 104.415    0.321
#> 3         jahrgang                       12 : 13 0.240 3.306  90.538    0.313
#> 4           geschl           Männlich : Weiblich 0.178 4.106 247.426    0.537
#> 5  jahrgang:geschl 11 & Männlich : 11 & Weiblich 0.285 3.918  97.980    0.408
#> 6  jahrgang:geschl 11 & Männlich : 12 & Männlich 0.268 1.950  96.767   -0.363
#> 7  jahrgang:geschl 11 & Männlich : 12 & Weiblich 0.276 2.410  97.724   -0.194
#> 8  jahrgang:geschl 11 & Männlich : 13 & Männlich 0.336 2.444  50.317   -0.246
#> 9  jahrgang:geschl 11 & Männlich : 13 & Weiblich 0.326 5.987  52.434    1.396
#> 10 jahrgang:geschl 11 & Weiblich : 12 & Männlich 0.270 2.196  96.445   -1.952
#> 11 jahrgang:geschl 11 & Weiblich : 12 & Weiblich 0.278 1.628  97.557   -1.783
#> 12 jahrgang:geschl 11 & Weiblich : 13 & Männlich 0.337 0.881  50.972   -1.832
#> 13 jahrgang:geschl 11 & Weiblich : 13 & Weiblich 0.328 2.546  53.113   -0.190
#> 14 jahrgang:geschl 12 & Männlich : 12 & Weiblich 0.261 0.542  97.648   -0.872
#> 15 jahrgang:geschl 12 & Männlich : 13 & Männlich 0.323 0.919  45.471   -0.940
#> 16 jahrgang:geschl 12 & Männlich : 13 & Weiblich 0.313 4.559  47.347    0.704
#> 17 jahrgang:geschl 12 & Weiblich : 13 & Männlich 0.329 0.472  47.962   -1.163
#> 18 jahrgang:geschl 12 & Weiblich : 13 & Weiblich 0.320 4.025  49.976    0.480
#> 19 jahrgang:geschl 13 & Männlich : 13 & Weiblich 0.372 3.037  47.900    0.037
#>    conf_high adj.p.value
#> 1      0.719       0.983
#> 2      2.019       0.004
#> 3      1.927       0.004
#> 4      1.527       0.000
#> 5      2.752       0.002
#> 6      1.843       0.379
#> 7      2.074       0.163
#> 8      2.566       0.161
#> 9      4.124       0.000
#> 10     0.272       0.249
#> 11     0.503       0.582
#> 12     0.992       0.949
#> 13     2.550       0.129
#> 14     1.272       0.994
#> 15     1.780       0.940
#> 16     3.336       0.000
#> 17     1.603       0.997
#> 18     3.160       0.003
#> 19     3.163       0.042