Extended Pairwise t-Tests

This is an extension of pairwise.t.test that's meant to deal with interactions our of the box, while also performing pairwise tests for the primary terms. The output of the function is modeled after TukeyHSD, unfortunately without confidence intervals or test statistic though.

tadaa_pairwise_t(data, response, group1, group2 = NULL, p.adjust = "bonf",
  paired = FALSE, pool.sd = !paired, alternative = "two.sided",
  print = "df")

Arguments

data
A data.frame containing the variables.

response
The response variable, i.e. the dependent numeric vector.

group1
The grouping variables, typically a factor.

group2
(Optional) second grouping variable.

p.adjust
The p-adjustment method, see p.adjust.methods, passed to pairwise.t.test. Additionally, sidak supported as a method not supported by p.adjust, as is sidakSD for the Sidak step-down procedure.

paired
Defaults to FALSE, also passed to pairwise.t.test.

pool.sd
Defaults to the inverse of paired, passed to pairwise.t.test.

alternative
Defaults to two.sided, also passed to pairwise.t.test.

print
Print method, defaults to df for data.frame output, otherwise passed to sprinkle_print_method.

Value

A data.frame with columns term, comparison and adj.p.value.

Note

The adjustment method is applied within each term, meaning that the number of pairwise t-tests counted for the adjustment is only equal to the number of rows per term of the output. The additional Sidak adjustment method uses the following method: p_adj <- 1 - pbinom(q = 0, size = length(p_values), prob = p_values) And is sometimes preferred over Bonferroni. The Sidak-like (1987) step-down procedure (sidakSD) is an improvement over the Holm's (1979) step-down procedure.

References

https://stats.stackexchange.com/questions/20825/sidak-or-bonferroni https://rdrr.io/rforge/mutoss/man/SidakSD.html

See also

tadaa_pairwise_tukey, tadaa_pairwise_gh

Other Tadaa-functions: tadaa_aov, tadaa_chisq, tadaa_kruskal, tadaa_levene, tadaa_nom, tadaa_normtest, tadaa_one_sample, tadaa_ord, tadaa_pairwise_gh, tadaa_pairwise_tukey, tadaa_t.test, tadaa_wilcoxon

Examples

tadaa_pairwise_t(ngo, deutsch, jahrgang, geschl, p.adjust = "none", print = "console")

#>               term                    comparison p (adj.)
#> 1         jahrgang                       12 - 11     0.86
#> 2         jahrgang                       13 - 11  < 0.001
#> 3         jahrgang                       13 - 12    0.001
#> 4           geschl           Weiblich - Männlich  < 0.001
#> 5  jahrgang:geschl 12 & Männlich - 11 & Männlich    0.055
#> 6  jahrgang:geschl 13 & Männlich - 11 & Männlich    0.014
#> 7  jahrgang:geschl 11 & Weiblich - 11 & Männlich  < 0.001
#> 8  jahrgang:geschl 12 & Weiblich - 11 & Männlich    0.015
#> 9  jahrgang:geschl 13 & Weiblich - 11 & Männlich  < 0.001
#> 10 jahrgang:geschl 13 & Männlich - 12 & Männlich     0.37
#> 11 jahrgang:geschl 11 & Weiblich - 12 & Männlich     0.03
#> 12 jahrgang:geschl 12 & Weiblich - 12 & Männlich      0.6
#> 13 jahrgang:geschl 13 & Weiblich - 12 & Männlich  < 0.001
#> 14 jahrgang:geschl 11 & Weiblich - 13 & Männlich     0.37
#> 15 jahrgang:geschl 12 & Weiblich - 13 & Männlich     0.64
#> 16 jahrgang:geschl 13 & Weiblich - 13 & Männlich    0.003
#> 17 jahrgang:geschl 12 & Weiblich - 11 & Weiblich    0.097
#> 18 jahrgang:geschl 13 & Weiblich - 11 & Weiblich    0.013
#> 19 jahrgang:geschl 13 & Weiblich - 12 & Weiblich  < 0.001
#> 
#> 

tadaa_pairwise_t(ngo, deutsch, jahrgang, geschl, p.adjust = "bonf", print = "console")

#>               term                    comparison p (adj.)
#> 1         jahrgang                       12 - 11   > 0.99
#> 2         jahrgang                       13 - 11    0.003
#> 3         jahrgang                       13 - 12    0.004
#> 4           geschl           Weiblich - Männlich  < 0.001
#> 5  jahrgang:geschl 12 & Männlich - 11 & Männlich     0.82
#> 6  jahrgang:geschl 13 & Männlich - 11 & Männlich     0.21
#> 7  jahrgang:geschl 11 & Weiblich - 11 & Männlich  < 0.001
#> 8  jahrgang:geschl 12 & Weiblich - 11 & Männlich     0.22
#> 9  jahrgang:geschl 13 & Weiblich - 11 & Männlich  < 0.001
#> 10 jahrgang:geschl 13 & Männlich - 12 & Männlich   > 0.99
#> 11 jahrgang:geschl 11 & Weiblich - 12 & Männlich     0.44
#> 12 jahrgang:geschl 12 & Weiblich - 12 & Männlich   > 0.99
#> 13 jahrgang:geschl 13 & Weiblich - 12 & Männlich  < 0.001
#> 14 jahrgang:geschl 11 & Weiblich - 13 & Männlich   > 0.99
#> 15 jahrgang:geschl 12 & Weiblich - 13 & Männlich   > 0.99
#> 16 jahrgang:geschl 13 & Weiblich - 13 & Männlich    0.052
#> 17 jahrgang:geschl 12 & Weiblich - 11 & Weiblich   > 0.99
#> 18 jahrgang:geschl 13 & Weiblich - 11 & Weiblich     0.19
#> 19 jahrgang:geschl 13 & Weiblich - 12 & Weiblich    0.002
#> 
#> 

tadaa_pairwise_t(ngo, deutsch, jahrgang, geschl, p.adjust = "sidak", print = "console")

#>               term                    comparison p (adj.)
#> 1         jahrgang                       12 - 11   > 0.99
#> 2         jahrgang                       13 - 11    0.003
#> 3         jahrgang                       13 - 12    0.004
#> 4           geschl           Weiblich - Männlich  < 0.001
#> 5  jahrgang:geschl 12 & Männlich - 11 & Männlich     0.57
#> 6  jahrgang:geschl 13 & Männlich - 11 & Männlich     0.19
#> 7  jahrgang:geschl 11 & Weiblich - 11 & Männlich  < 0.001
#> 8  jahrgang:geschl 12 & Weiblich - 11 & Männlich      0.2
#> 9  jahrgang:geschl 13 & Weiblich - 11 & Männlich  < 0.001
#> 10 jahrgang:geschl 13 & Männlich - 12 & Männlich   > 0.99
#> 11 jahrgang:geschl 11 & Weiblich - 12 & Männlich     0.36
#> 12 jahrgang:geschl 12 & Weiblich - 12 & Männlich   > 0.99
#> 13 jahrgang:geschl 13 & Weiblich - 12 & Männlich  < 0.001
#> 14 jahrgang:geschl 11 & Weiblich - 13 & Männlich   > 0.99
#> 15 jahrgang:geschl 12 & Weiblich - 13 & Männlich   > 0.99
#> 16 jahrgang:geschl 13 & Weiblich - 13 & Männlich    0.051
#> 17 jahrgang:geschl 12 & Weiblich - 11 & Weiblich     0.78
#> 18 jahrgang:geschl 13 & Weiblich - 11 & Weiblich     0.17
#> 19 jahrgang:geschl 13 & Weiblich - 12 & Weiblich    0.002
#> 
#>