Tukey HSD pairwise comparisons

This function is merely a thin wrapper around stats::TukeyHSD with tidying done by broom::tidy and optional formatting via pixiedust::sprinkle. Its input is not a aov model like in the original function, but instead the aov model is fit internally based on the arguments given. This is meant to enable a consistent usage between the tadaa_pairwise-functions.

tadaa_pairwise_tukey(data, response, group1, group2 = NULL, print = "df",
  ...)

Arguments

data
A data.frame containing the variables.

response
The response variable, i.e. the dependent numeric vector.

group1
The grouping variables, typically a factor.

group2
(Optional) second grouping variable.

print
Print method, defaults to df for data.frame output, otherwise passed to pixiedust::sprinkle_print_method.

...
Further arguments passed to stats::TukeyHSD

Value

A data.frame or pixiedust::dust object depending on print.

See also

tadaa_pairwise_t(), tadaa_pairwise_gh()

Other Tadaa-functions: tadaa_aov, tadaa_chisq, tadaa_kruskal, tadaa_levene, tadaa_nom, tadaa_normtest, tadaa_one_sample, tadaa_ord, tadaa_pairwise_gh, tadaa_pairwise_t, tadaa_t.test, tadaa_wilcoxon

Examples

tadaa_pairwise_tukey(data = ngo, deutsch, jahrgang, geschl)

#>               term              comparison estimate     conf.low  conf.high
#> 1         jahrgang                   12-11    0.050 -0.589702593 0.68970259
#> 2         jahrgang                   13-11    1.170  0.386527530 1.95347247
#> 3         jahrgang                   13-12    1.120  0.336527530 1.90347247
#> 4           geschl       Weiblich-Männlich    1.032  0.554066695 1.50993331
#> 5  jahrgang:geschl 12:Männlich-11:Männlich    0.740 -0.362023980 1.84202398
#> 6  jahrgang:geschl 13:Männlich-11:Männlich    1.160 -0.189698218 2.50969822
#> 7  jahrgang:geschl 11:Weiblich-11:Männlich    1.580  0.477976020 2.68202398
#> 8  jahrgang:geschl 12:Weiblich-11:Männlich    0.940 -0.162023980 2.04202398
#> 9  jahrgang:geschl 13:Weiblich-11:Männlich    2.760  1.410301782 4.10969822
#> 10 jahrgang:geschl 13:Männlich-12:Männlich    0.420 -0.929698218 1.76969822
#> 11 jahrgang:geschl 11:Weiblich-12:Männlich    0.840 -0.262023980 1.94202398
#> 12 jahrgang:geschl 12:Weiblich-12:Männlich    0.200 -0.902023980 1.30202398
#> 13 jahrgang:geschl 13:Weiblich-12:Männlich    2.020  0.670301782 3.36969822
#> 14 jahrgang:geschl 11:Weiblich-13:Männlich    0.420 -0.929698218 1.76969822
#> 15 jahrgang:geschl 12:Weiblich-13:Männlich   -0.220 -1.569698218 1.12969822
#> 16 jahrgang:geschl 13:Weiblich-13:Männlich    1.600  0.041502741 3.15849726
#> 17 jahrgang:geschl 12:Weiblich-11:Weiblich   -0.640 -1.742023980 0.46202398
#> 18 jahrgang:geschl 13:Weiblich-11:Weiblich    1.180 -0.169698218 2.52969822
#> 19 jahrgang:geschl 13:Weiblich-12:Weiblich    1.820  0.470301782 3.16969822
#>      adj.p.value
#> 1  9.8144775e-01
#> 2  1.4811944e-03
#> 3  2.4998298e-03
#> 4  3.0038898e-05
#> 5  3.8706989e-01
#> 6  1.3758052e-01
#> 7  7.3634706e-04
#> 8  1.4346679e-01
#> 9  2.0684748e-07
#> 10 9.4771171e-01
#> 11 2.4641083e-01
#> 12 9.9529981e-01
#> 13 3.5420602e-04
#> 14 9.4771171e-01
#> 15 9.9717156e-01
#> 16 4.0403443e-02
#> 17 5.5416077e-01
#> 18 1.2480164e-01
#> 19 1.8979635e-03

tadaa_pairwise_tukey(data = ngo, deutsch, jahrgang, print = "console")

#>       Term Comparison Diff CI_low CI_high p (adj.)
#> 1 jahrgang      12-11 0.05 -0.619   0.719     0.98
#> 2 jahrgang      13-11 1.17  0.351   1.989   < 0.01
#> 3 jahrgang      13-12 1.12  0.301   1.939   < 0.01
#> 
#>