Tadaa, one-sample tests!

If sigma is omitted, the function will just perform a one-sample stats::t.test, but if sigma is provided, a z-test is performed. It basically works the same way, except that we pretend we know the population sigma and use the normal distribution for comparison.

tadaa_one_sample(data = NULL, x, mu, sigma = NULL,
  direction = "two.sided", na.rm = FALSE, conf.level = 0.95,
  print = c("df", "console", "html", "markdown"))

Arguments

data
A data.frame (optional).

x
A numeric vector or bare column name of data.

mu
The true mean (\(\mu\)) to test for.

sigma
Population sigma. If supplied, a z-test is performed, otherwise a one-sample stats::t.test is performed.

direction
Test direction, like alternative in t.test.

na.rm
Whether to drop NA values. Default is FALSE.

conf.level
Confidence level used for power and CI, default is 0.95.

print
Print method, default df: A regular data.frame. Otherwise passed to pixiedust::sprinkle_print_method for fancyness.

Value

A data.frame by default, otherwise dust object, depending on print.

See also

Examples

set.seed(42)
df <- data.frame(x = rnorm(n = 20, mean = 100, sd = 1))

tadaa_one_sample(df, x, mu = 101, sigma = 1)

#>   estimate statistic        se      p.value conf.low conf.high method
#> 1 100.1919 -3.613844 0.2236068 0.0003016912 70.61548  129.7684 z-Test
#>   alternative        d    power
#> 1   two.sided -0.80808 0.950924

# No data.frame, just a vector
tadaa_one_sample(x = rnorm(20), mu = 0)

#> # A tibble: 1 x 11
#>   estimate statistic p.value parameter conf.low conf.high method alternative
#>      <dbl>     <dbl>   <dbl>     <dbl>    <dbl>     <dbl> <chr>  <chr>      
#> 1   -0.271     -1.09   0.288        19   -0.790     0.248 One S… two.sided  
#> # … with 3 more variables: se <dbl>, d <dbl>, power <dbl>