Extended Pairwise t-Tests

This is an extension of stats::pairwise.t.test that's meant to deal with interactions our of the box, while also performing pairwise tests for the primary terms. The output of the function is modeled after stats::TukeyHSD, unfortunately without confidence intervals or test statistic though.

tadaa_pairwise_t(
  data,
  response,
  group1,
  group2 = NULL,
  p.adjust = "bonf",
  paired = FALSE,
  pool.sd = !paired,
  alternative = "two.sided",
  print = "df"
)

Arguments

data
A data.frame containing the variables.

response
The response variable, i.e. the dependent numeric vector.

group1
The grouping variables, typically a factor.

group2
(Optional) second grouping variable.

p.adjust
The p-adjustment method, see stats::p.adjust.methods, passed to stats::pairwise.t.test. Additionally, sidak is supported as a method, which is not the case with stats::p.adjust, as is sidakSD for the Sidak step-down procedure.

paired
Defaults to FALSE, also passed to stats::pairwise.t.test.

pool.sd
Defaults to the inverse of paired, passed to stats::pairwise.t.test.

alternative
Defaults to two.sided, also passed to stats::pairwise.t.test.

print
Print method, defaults to df for data.frame output, otherwise passed to pixiedust::sprinkle_print_method.

Value

A data.frame with columns term, comparison and adj.p.value.

Note

The adjustment method is applied within each term, meaning that the number of pairwise t-tests counted for the adjustment is only equal to the number of rows per term of the output. The additional Sidak adjustment method uses the following method: p_adj <- 1 - pbinom(q = 0, size = length(p_values), prob = p_values) And is sometimes preferred over Bonferroni. The Sidak-like (1987) step-down procedure (sidakSD) is an improvement over the Holm's (1979) step-down procedure.

References

https://stats.stackexchange.com/questions/20825/sidak-or-bonferroni

https://rdrr.io/rforge/mutoss/man/SidakSD.html

See also

Examples

tadaa_pairwise_t(ngo, deutsch, jahrgang, geschl, p.adjust = "none", print = "console")

#>               term                    comparison p (adj.)
#> 1         jahrgang                       12 - 11      .86
#> 2         jahrgang                       13 - 11   < .001
#> 3         jahrgang                       13 - 12    < .01
#> 4           geschl           Weiblich - Männlich   < .001
#> 5  jahrgang:geschl 12 & Männlich - 11 & Männlich     .055
#> 6  jahrgang:geschl 13 & Männlich - 11 & Männlich    < .05
#> 7  jahrgang:geschl 11 & Weiblich - 11 & Männlich   < .001
#> 8  jahrgang:geschl 12 & Weiblich - 11 & Männlich    < .05
#> 9  jahrgang:geschl 13 & Weiblich - 11 & Männlich   < .001
#> 10 jahrgang:geschl 13 & Männlich - 12 & Männlich     .372
#> 11 jahrgang:geschl 11 & Weiblich - 12 & Männlich    < .05
#> 12 jahrgang:geschl 12 & Weiblich - 12 & Männlich     .603
#> 13 jahrgang:geschl 13 & Weiblich - 12 & Männlich   < .001
#> 14 jahrgang:geschl 11 & Weiblich - 13 & Männlich     .372
#> 15 jahrgang:geschl 12 & Weiblich - 13 & Männlich      .64
#> 16 jahrgang:geschl 13 & Weiblich - 13 & Männlich    < .01
#> 17 jahrgang:geschl 12 & Weiblich - 11 & Weiblich     .097
#> 18 jahrgang:geschl 13 & Weiblich - 11 & Weiblich    < .05
#> 19 jahrgang:geschl 13 & Weiblich - 12 & Weiblich   < .001
#> 
#> 

tadaa_pairwise_t(ngo, deutsch, jahrgang, geschl, p.adjust = "bonf", print = "console")

#>               term                    comparison p (adj.)
#> 1         jahrgang                       12 - 11    > .99
#> 2         jahrgang                       13 - 11    < .01
#> 3         jahrgang                       13 - 12    < .01
#> 4           geschl           Weiblich - Männlich   < .001
#> 5  jahrgang:geschl 12 & Männlich - 11 & Männlich     .824
#> 6  jahrgang:geschl 13 & Männlich - 11 & Männlich     .214
#> 7  jahrgang:geschl 11 & Weiblich - 11 & Männlich   < .001
#> 8  jahrgang:geschl 12 & Weiblich - 11 & Männlich     .225
#> 9  jahrgang:geschl 13 & Weiblich - 11 & Männlich   < .001
#> 10 jahrgang:geschl 13 & Männlich - 12 & Männlich    > .99
#> 11 jahrgang:geschl 11 & Weiblich - 12 & Männlich     .443
#> 12 jahrgang:geschl 12 & Weiblich - 12 & Männlich    > .99
#> 13 jahrgang:geschl 13 & Weiblich - 12 & Männlich   < .001
#> 14 jahrgang:geschl 11 & Weiblich - 13 & Männlich    > .99
#> 15 jahrgang:geschl 12 & Weiblich - 13 & Männlich    > .99
#> 16 jahrgang:geschl 13 & Weiblich - 13 & Männlich     .052
#> 17 jahrgang:geschl 12 & Weiblich - 11 & Weiblich    > .99
#> 18 jahrgang:geschl 13 & Weiblich - 11 & Weiblich      .19
#> 19 jahrgang:geschl 13 & Weiblich - 12 & Weiblich    < .01
#> 
#> 

tadaa_pairwise_t(ngo, deutsch, jahrgang, geschl, p.adjust = "sidak", print = "console")

#>               term                    comparison p (adj.)
#> 1         jahrgang                       12 - 11    > .99
#> 2         jahrgang                       13 - 11    < .01
#> 3         jahrgang                       13 - 12    < .01
#> 4           geschl           Weiblich - Männlich   < .001
#> 5  jahrgang:geschl 12 & Männlich - 11 & Männlich     .571
#> 6  jahrgang:geschl 13 & Männlich - 11 & Männlich     .194
#> 7  jahrgang:geschl 11 & Weiblich - 11 & Männlich   < .001
#> 8  jahrgang:geschl 12 & Weiblich - 11 & Männlich     .203
#> 9  jahrgang:geschl 13 & Weiblich - 11 & Männlich   < .001
#> 10 jahrgang:geschl 13 & Männlich - 12 & Männlich    > .99
#> 11 jahrgang:geschl 11 & Weiblich - 12 & Männlich     .362
#> 12 jahrgang:geschl 12 & Weiblich - 12 & Männlich    > .99
#> 13 jahrgang:geschl 13 & Weiblich - 12 & Männlich   < .001
#> 14 jahrgang:geschl 11 & Weiblich - 13 & Männlich    > .99
#> 15 jahrgang:geschl 12 & Weiblich - 13 & Männlich    > .99
#> 16 jahrgang:geschl 13 & Weiblich - 13 & Männlich     .051
#> 17 jahrgang:geschl 12 & Weiblich - 11 & Weiblich     .782
#> 18 jahrgang:geschl 13 & Weiblich - 11 & Weiblich     .174
#> 19 jahrgang:geschl 13 & Weiblich - 12 & Weiblich    < .01
#> 
#>