Tukey HSD pairwise comparisons

This function is merely a thin wrapper around stats::TukeyHSD with tidying done by broom::tidy and optional formatting via pixiedust::sprinkle. Its input is not a aov model like in the original function, but instead the aov model is fit internally based on the arguments given. This is meant to enable a consistent usage between the tadaa_pairwise-functions.

tadaa_pairwise_tukey(data, response, group1, group2 = NULL,
  print = "df", ...)

Arguments

data
A data.frame containing the variables.

response
The response variable, i.e. the dependent numeric vector.

group1
The grouping variables, typically a factor.

group2
(Optional) second grouping variable.

print
Print method, defaults to df for data.frame output, otherwise passed to pixiedust::sprinkle_print_method.

...
Further arguments passed to stats::TukeyHSD

Value

A data.frame or pixiedust::dust object depending on print.

See also

Examples

tadaa_pairwise_tukey(data = ngo, deutsch, jahrgang, geschl)

#> # A tibble: 19 x 6
#>    term           comparison            estimate conf.low conf.high adj.p.value
#>    <chr>          <chr>                    <dbl>    <dbl>     <dbl>       <dbl>
#>  1 jahrgang       12-11                   0.0500  -0.590      0.690 0.981      
#>  2 jahrgang       13-11                   1.17     0.387      1.95  0.00148    
#>  3 jahrgang       13-12                   1.12     0.337      1.90  0.00250    
#>  4 geschl         Weiblich-Männlich       1.03     0.554      1.51  0.0000300  
#>  5 jahrgang:gesc… 12:Männlich-11:Männl…   0.740   -0.362      1.84  0.387      
#>  6 jahrgang:gesc… 13:Männlich-11:Männl…   1.16    -0.190      2.51  0.138      
#>  7 jahrgang:gesc… 11:Weiblich-11:Männl…   1.58     0.478      2.68  0.000736   
#>  8 jahrgang:gesc… 12:Weiblich-11:Männl…   0.940   -0.162      2.04  0.143      
#>  9 jahrgang:gesc… 13:Weiblich-11:Männl…   2.76     1.41       4.11  0.000000207
#> 10 jahrgang:gesc… 13:Männlich-12:Männl…   0.420   -0.930      1.77  0.948      
#> 11 jahrgang:gesc… 11:Weiblich-12:Männl…   0.840   -0.262      1.94  0.246      
#> 12 jahrgang:gesc… 12:Weiblich-12:Männl…   0.200   -0.902      1.30  0.995      
#> 13 jahrgang:gesc… 13:Weiblich-12:Männl…   2.02     0.670      3.37  0.000354   
#> 14 jahrgang:gesc… 11:Weiblich-13:Männl…   0.420   -0.930      1.77  0.948      
#> 15 jahrgang:gesc… 12:Weiblich-13:Männl…  -0.220   -1.57       1.13  0.997      
#> 16 jahrgang:gesc… 13:Weiblich-13:Männl…   1.60     0.0415     3.16  0.0404     
#> 17 jahrgang:gesc… 12:Weiblich-11:Weibl…  -0.640   -1.74       0.462 0.554      
#> 18 jahrgang:gesc… 13:Weiblich-11:Weibl…   1.18    -0.170      2.53  0.125      
#> 19 jahrgang:gesc… 13:Weiblich-12:Weibl…   1.82     0.470      3.17  0.00190    

tadaa_pairwise_tukey(data = ngo, deutsch, jahrgang, print = "console")

#>       Term Comparison Diff CI_low CI_high p (adj.)
#> 1 jahrgang      12-11 0.05 -0.619   0.719     .983
#> 2 jahrgang      13-11 1.17  0.351   1.989    < .01
#> 3 jahrgang      13-12 1.12  0.301   1.939    < .01
#> 
#>