Tukey HSD pairwise comparisons

This function is merely a thin wrapper around stats::TukeyHSD with tidying done by broom::tidy and optional formatting via pixiedust::sprinkle. Its input is not a aov model like in the original function, but instead the aov model is fit internally based on the arguments given. This is meant to enable a consistent usage between the tadaa_pairwise-functions.

tadaa_pairwise_tukey(data, response, group1, group2 = NULL, print = "df",
  ...)

Arguments

data
A data.frame containing the variables.

response
The response variable, i.e. the dependent numeric vector.

group1
The grouping variables, typically a factor.

group2
(Optional) second grouping variable.

print
Print method, defaults to df for data.frame output, otherwise passed to pixiedust::sprinkle_print_method.

...
Further arguments passed to stats::TukeyHSD

Value

A data.frame or pixiedust::dust object depending on print.

See also

Examples

tadaa_pairwise_tukey(data = ngo, deutsch, jahrgang, geschl)

#>               term              comparison estimate    conf.low conf.high
#> 1         jahrgang                   12-11    0.050 -0.58970259 0.6897026
#> 2         jahrgang                   13-11    1.170  0.38652753 1.9534725
#> 3         jahrgang                   13-12    1.120  0.33652753 1.9034725
#> 4           geschl       Weiblich-Männlich    1.032  0.55406669 1.5099333
#> 5  jahrgang:geschl 12:Männlich-11:Männlich    0.740 -0.36202398 1.8420240
#> 6  jahrgang:geschl 13:Männlich-11:Männlich    1.160 -0.18969822 2.5096982
#> 7  jahrgang:geschl 11:Weiblich-11:Männlich    1.580  0.47797602 2.6820240
#> 8  jahrgang:geschl 12:Weiblich-11:Männlich    0.940 -0.16202398 2.0420240
#> 9  jahrgang:geschl 13:Weiblich-11:Männlich    2.760  1.41030178 4.1096982
#> 10 jahrgang:geschl 13:Männlich-12:Männlich    0.420 -0.92969822 1.7696982
#> 11 jahrgang:geschl 11:Weiblich-12:Männlich    0.840 -0.26202398 1.9420240
#> 12 jahrgang:geschl 12:Weiblich-12:Männlich    0.200 -0.90202398 1.3020240
#> 13 jahrgang:geschl 13:Weiblich-12:Männlich    2.020  0.67030178 3.3696982
#> 14 jahrgang:geschl 11:Weiblich-13:Männlich    0.420 -0.92969822 1.7696982
#> 15 jahrgang:geschl 12:Weiblich-13:Männlich   -0.220 -1.56969822 1.1296982
#> 16 jahrgang:geschl 13:Weiblich-13:Männlich    1.600  0.04150274 3.1584973
#> 17 jahrgang:geschl 12:Weiblich-11:Weiblich   -0.640 -1.74202398 0.4620240
#> 18 jahrgang:geschl 13:Weiblich-11:Weiblich    1.180 -0.16969822 2.5296982
#> 19 jahrgang:geschl 13:Weiblich-12:Weiblich    1.820  0.47030178 3.1696982
#>     adj.p.value
#> 1  9.814478e-01
#> 2  1.481194e-03
#> 3  2.499830e-03
#> 4  3.003890e-05
#> 5  3.870699e-01
#> 6  1.375805e-01
#> 7  7.363471e-04
#> 8  1.434668e-01
#> 9  2.068475e-07
#> 10 9.477117e-01
#> 11 2.464108e-01
#> 12 9.952998e-01
#> 13 3.542060e-04
#> 14 9.477117e-01
#> 15 9.971716e-01
#> 16 4.040344e-02
#> 17 5.541608e-01
#> 18 1.248016e-01
#> 19 1.897963e-03

tadaa_pairwise_tukey(data = ngo, deutsch, jahrgang, print = "console")

#>       Term Comparison Diff CI_low CI_high p (adj.)
#> 1 jahrgang      12-11 0.05 -0.619   0.719     0.98
#> 2 jahrgang      13-11 1.17  0.351   1.989   < 0.01
#> 3 jahrgang      13-12 1.12  0.301   1.939   < 0.01
#> 
#>