Tukey HSD pairwise comparisons

This function is merely a thin wrapper around TukeyHSD with tidying done by tidy and optional formatting via sprinkle. Its input is not a aov model like in the original function, but instead the aov model is fit internally based on the arguments given. This is meant to enable a consistent usage between the tadaa_pairwise-functions.

tadaa_pairwise_tukey(data, response, group1, group2 = NULL, print = "df",
  ...)

Arguments

data
A data.frame containing the variables.

response
The response variable, i.e. the dependent numeric vector.

group1
The grouping variables, typically a factor.

group2
(Optional) second grouping variable.

print
Print method, defaults to df for data.frame output, otherwise passed to sprinkle_print_method.

...
Further arguments passed to TukeyHSD

Value

A data.frame or dust object depending on print.

See also

tadaa_pairwise_t, tadaa_pairwise_gh

Other Tadaa-functions: tadaa_aov, tadaa_chisq, tadaa_kruskal, tadaa_levene, tadaa_nom, tadaa_normtest, tadaa_one_sample, tadaa_ord, tadaa_pairwise_gh, tadaa_pairwise_t, tadaa_t.test, tadaa_wilcoxon

Examples

tadaa_pairwise_tukey(data = ngo, deutsch, jahrgang, geschl)

#>               term              comparison estimate         conf.low
#> 1         jahrgang                   12-11    0.050 -0.5897025931261
#> 2         jahrgang                   13-11    1.170  0.3865275298529
#> 3         jahrgang                   13-12    1.120  0.3365275298529
#> 4           geschl       Weiblich-Männlich    1.032  0.5540666948627
#> 5  jahrgang:geschl 12:Männlich-11:Männlich    0.740 -0.3620239800300
#> 6  jahrgang:geschl 13:Männlich-11:Männlich    1.160 -0.1896982176922
#> 7  jahrgang:geschl 11:Weiblich-11:Männlich    1.580  0.4779760199700
#> 8  jahrgang:geschl 12:Weiblich-11:Männlich    0.940 -0.1620239800300
#> 9  jahrgang:geschl 13:Weiblich-11:Männlich    2.760  1.4103017823078
#> 10 jahrgang:geschl 13:Männlich-12:Männlich    0.420 -0.9296982176922
#> 11 jahrgang:geschl 11:Weiblich-12:Männlich    0.840 -0.2620239800300
#> 12 jahrgang:geschl 12:Weiblich-12:Männlich    0.200 -0.9020239800300
#> 13 jahrgang:geschl 13:Weiblich-12:Männlich    2.020  0.6703017823078
#> 14 jahrgang:geschl 11:Weiblich-13:Männlich    0.420 -0.9296982176922
#> 15 jahrgang:geschl 12:Weiblich-13:Männlich   -0.220 -1.5696982176922
#> 16 jahrgang:geschl 13:Weiblich-13:Männlich    1.600  0.0415027413813
#> 17 jahrgang:geschl 12:Weiblich-11:Weiblich   -0.640 -1.7420239800300
#> 18 jahrgang:geschl 13:Weiblich-11:Weiblich    1.180 -0.1696982176922
#> 19 jahrgang:geschl 13:Weiblich-12:Weiblich    1.820  0.4703017823078
#>         conf.high       adj.p.value
#> 1  0.689702593126 9.81447753429e-01
#> 2  1.953472470147 1.48119440363e-03
#> 3  1.903472470147 2.49982979842e-03
#> 4  1.509933305137 3.00388975241e-05
#> 5  1.842023980030 3.87069885176e-01
#> 6  2.509698217692 1.37580520957e-01
#> 7  2.682023980030 7.36347059978e-04
#> 8  2.042023980030 1.43466787839e-01
#> 9  4.109698217692 2.06847479922e-07
#> 10 1.769698217692 9.47711705732e-01
#> 11 1.942023980030 2.46410825000e-01
#> 12 1.302023980030 9.95299813499e-01
#> 13 3.369698217692 3.54206016988e-04
#> 14 1.769698217692 9.47711705732e-01
#> 15 1.129698217692 9.97171561397e-01
#> 16 3.158497258619 4.04034427567e-02
#> 17 0.462023980030 5.54160770911e-01
#> 18 2.529698217692 1.24801639876e-01
#> 19 3.169698217692 1.89796345526e-03

tadaa_pairwise_tukey(data = ngo, deutsch, jahrgang, print = "console")

#> Warning: attributes are not identical across measure variables;
#> they will be dropped

#>       Term Comparison Diff CI_low CI_high p (adj.)
#> 1 jahrgang      12-11 0.05 -0.619   0.719     0.98
#> 2 jahrgang      13-11 1.17  0.351   1.989    0.003
#> 3 jahrgang      13-12 1.12  0.301   1.939    0.004
#> 
#>